style: changing header buttons to anchors
